requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.0.0
python-decouple==3.8
django-filter==23.3
